src/main/java/mh/springboot/model/Entity.java
package mh.springboot.model;

import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;

@JsonDeserialize(builder = Entity.Builder.class)
public class Entity {

    private Long id;
    private String name;

    public Entity(Builder builder) {
        this.id = builder.id;
        this.name = builder.name;
    }

    @JsonPOJOBuilder
    public static class Builder {
        private Long id;
        private String name;

        public Builder withId(Long id) {
            this.id = id;
            return this;
        }

        public Builder withName(String name) {
            this.name = name;
            return this;
        }

        public Entity build() {
            return new Entity(this);
        }
    }

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(final String name) {
        this.name = name;
    }
}
